requirements.txt
# local package
-e .

# external requirements
click
seaborn
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
pyngrok
streamlit
sklearn
# easyocr